+++ app/(route)/activity/register/_components/TextEditor/index.tsx
@@ -7,7 +7,7 @@ import type { Dispatch, SetStateAction } from 'react';
 import { useEffect, useMemo, useState } from 'react';
 import dynamic from 'next/dynamic';
 
-import type { Activity, ActivityEdit } from '@/_types/activities/formTypes';
+import type { Activity, ActivityEdit } from '@/_types/activities/form.types';
 
 import Textarea from '@/_components/Textarea';
